chore(TierPage): Move longDescription to a separate component
